fix + test: Investigate minor bugs and test for policy action sampling. Works.

+++ ast_sac/env_wrapper/normalized_box_env.py
@@ -30,8 +30,8 @@ class NormalizedBoxEnv(ProxyEnv):
         self._reward_scale = reward_scale
         self._obs_mean = obs_mean
         self._obs_std = obs_std
-        ub = np.ones(self._wrapped_env.action_space.shape)
-        self.action_space = Box(-1 * ub, ub)
+        ub = np.ones(self._wrapped_env.action_space.shape, dtype=np.float32)
+        self.action_space = Box(-1 * ub, ub, dtype=np.float32)
 
     def estimate_obs_stats(self, obs_batch, override_values=False):
         if self._obs_mean is not None and not override_values:
